requirements.txt
flask==2.3.3
Flask-SQLAlchemy==3.1.1
slack-sdk==3.25.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
pytest==7.4.2
requests==2.31.0
click==8.1.7
websocket-client==1.6.1 